Use CIDR notation for client routes during wizard setup
Previously, the kit requested routes in IP - netmask notation, however,
a bug prevented the YAML from being properly assembled due to spaces in
the notation. Rather than attempt to fix that bug specifically, it was
decided that moving to CIDR was better option as it was already a
requested feature (#9)

This commit does NOT change the configuration YAML (this still uses
netmask format, as is required by OpenVPN). This commit only adds the
ability for the wizard to convert CIDR to the appropriate netmask.